Inventory Item complete, work order limitation added

+++ src/main/java/com/nextgenmanager/nextgenmanager/marketing/enquiry/service/EnquiryService.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 
 public interface EnquiryService {
 
-    public Enquiry getEnquiry(int id);
+    public Enquiry getEnquiry(Long id);
 
     public Page<EnquiryTableDTO> getAllActiveEnquiry(int page, int size,String sortBy, String sortDir, String enqNo, String companyName, LocalDate lastContactedDate,
                                                      LocalDate enqDate, LocalDate closedDate, Integer daysForNetFollowUp,
@@ -17,13 +17,13 @@ public interface EnquiryService {
                                                      String dateComparisonTypeClosedDate);
 
     public Page<Enquiry> getAllEnquiry(int page, int size,String sortBy, String sortDir);
-    public Enquiry updateEnquiry(Enquiry updatedEnquiry,int id);
+    public Enquiry updateEnquiry(Enquiry updatedEnquiry, Long id);
 
     public Enquiry createEnquiry(Enquiry newEnquiry);
 
-    public void deleteEnquiry(int id);
+    public void deleteEnquiry(Long id);
 
-    public void closeEnquiry(int id, String closeReason);
+    public void closeEnquiry(Long id, String closeReason);
 
     public Enquiry getEnquiryByEnquiryNo(String enquiryNo);
